remove AllButton from story, put it in the docs
