+++ src/components/Button/Button.mdx
@@ -1,0 +1,6 @@
+import { ButtonDocs } from './ButtonDocs';
+import { Unstyled } from '@storybook/blocks';
+
+<Unstyled>
+  <ButtonDocs />
+</Unstyled>
